Enh(Message): Move reactions to an independent component
